Add unite.vim and vimproc.vim
Use unite.vim for Ack.vim and ctrl-p replacement.  Unite can work with
standard :vimgrep/:grep, and using the changes in the .vimrc.d/unite.vim
file, use ack if it is available.  This can also be used for many other
list fuzzy searching purposes, so this might replace NERDTree if I so
choose in the future.

This does require `vimproc.vim`, which does have a C file that needs to
be compiled, but it is able to be compiled on all operating systems, so
I am fine with including this.  If needed, I can fallback to :vimgrep,
which is fine in a pinch.

I have also found that there are situations where using a different
search is better over others.  When in a massive git project, using
`git grep` tends to be quicker as it skips log and tmp dirs, and
anything not checked into the repo (usually what I want).  Ack is also
just faster overall, and will do log files by default.  This is what
ToggleUniteSearchCommand will do, and it allows me to switch the search
provider on the fly, defaulting to the initial grep provided by unite.