Add netrbrowsex workaround for OSX
When using vim-fugitive, the :Gbrowse command doesn't function properly
on OSX.  This is because th netrw plugin (ships with vim) doesn't
properly detect the OSX version of vim installed on Macs by default
(when using the `has("mac")` function in vimscript), which it would
normally set the program to the OSX `open` command.

Using this workaround, we override the netrw default to use `open` by
setting the `g:netrw_browsex_viewer` variable if the system is a unix
system (so it will respond to the `uname` system command), and it
includes "Darwin" in the result of that command.