Adds explaination and AirlineToggleBranch to airline config
The explaination is here just to further explain why these extra lines
exist in this file that do almost nothing.

Also added an AirlineToggleBranch function which is used to remove the
branch name from the status.  Useful when using long branch names and it
gets in the way of the filename and other relavent parts of the status
line.