Add shebang for mux command
Not sure why OSX keeps throwing up with this one, but this seemed to fix
the issue I was having where I was getting this:

    $ mux
    -bash(40855,0x7fff753d4000) malloc: *** error for object 0xb00007fade1802f3: pointer being freed was not allocated
    *** set a breakpoint in malloc_error_break to debug
    Abort trap: 6

Not sure.  The googles say it is a permission issue, but can't find a
difference between the permissions for `/bin/sh` and `/bin/bash`.  Oh
well, I give up.