Move 'autocmd QuickFixCmdPost' to search.vim
These for some reason lived in .vimrc.d/unite.vim, but really is more
relevant in .vimrc.d/search.vim, since it doesn't just affect unite, and
is more search related.

The search.vim file will also be expanded a bit, in the future, to be
more generic in what defines "search".