[tmux.conf] Update and re-arrange keybindings
This is in done for the config changes when updating from:

    tmux `2.3_3` -> `2.7`

Since the syntax changed for handling keybindings for copy mode, among
other things.  Didn't fully read the documentation on it, but this
should now support the same functionality I had, while also being
compatible with the new syntax.

Also moved a few things that were in the man `.tmux.conf` into the
`.tmux/keybindings.conf` file, since they were better suited there.