[.bashrc.d/.aliases] Remove rake alias
The last commit was kind of for show, but I really don't like this alias
any more.

It isn't that hard to run `bundle exec rake` when needed, or to make
sure your rake tasks are using a set of features that is a little more
backwards compatible, and the removal of being able to use the `rake`
bin from rubygems explicitly is kind of getting annoying, and the
convenience of this has lost it's luster.

As for the previous commit, still want it in my history for posterity,
but don't really feel I will add it back.