[.vim] Support `hub pull-request` syntax
As of this PR:  https://github.com/github/hub/pull/1939

There is support for syntax highlighting for pull request files when
using `hub pull-request`.

This change adds the `/usr/local/share/vim/vimfiles` directory as part
of pathogen's path structure, if it exists.  Other platforms could be
added in the future if desired.