Add .vim_plugin.sh
A helper script for adding `git submodules` to this repo.

Not part of `bin/` since this isn't something that should be used
outside of this repo/directory, and shouldn't be available on the rest
of the system.

Simply there because I forget how to add these every time I come into
add a new one every half a year...