[bin/fork] Skip fork step if already done
Uses `curl` to do a check to see if the repo has already been forked by
simply hitting the repo for the newly created repo.

If a 404 is returned for the $NEW_FORK_URL, it can then be assumed that
the repo doesn't exist under this user, so we are safe to execute the
fork API call for the given repo.

Otherwise, doing the clone afterwards works just fine, and the rest
ofthe steps now are still valid for setting up the new clone.