[.vimrc] Fix <leader> keybinds
When moving some common keybinds to a deticated keybindings.vim file,
the `let mapleader = ";"` line was also moved.

This proved to be a problem if it wasn't set and a keybinding made use
of it via `<leader>`.  This puts that line back in the `.vimrc`, and a
friendly reminder to not be a derp.