[.vimrc] Remove unite.vim
I don't use it, and it requires an extra plugin that requires
compilation, and I just can't be bothered.

Might switch to ctrl-p since that is still maintained, but would need to
feel the need to have a fuzzy finder again before I consider adding
that.  The :grep and :Ggrep commands have been working just fine for my
needs, and is mostly what I use anyway.