[.irbrc] Twilight code theme
This is a bit of a hack and probably only partially working, but this
changes the default color "theme" for `irb` when highlighting the code
to "twilight-esque".

The only change, besides monkey patching some parts of a constant, is
that the `dispatch_seq` needed to be changed since it hard codes a value
for symbols/constants, so this updates it to use the new colors that we
have defined, instead of what was hard coded.