bin/fans: allow for CUSTOM_SMC logic
The interface from the "custom smc" bin that I use from here:

https://github.com/hholtmann/smcFanControl/pull/108

Which allows this program to run on newer versions of OSX, has changed.
The "homebrew" version of this program from smcFanControl is out of
date, but can be used for older versions, and when I custom build this
version, I need a different set of flags passed.

This change allows supporting both.